requirements.txt
pandas==1.4.3
Pillow==9.4.0
streamlit==1.20.0
